# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ProjTrack - Project Submission Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export type LoginBodyRole = (typeof LoginBodyRole)[keyof typeof LoginBodyRole];

export const LoginBodyRole = {
  admin: "admin",
  teacher: "teacher",
  student: "student",
} as const;

export interface LoginBody {
  /** Email or ID */
  identifier: string;
  password: string;
  role: LoginBodyRole;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  teacher: "teacher",
  student: "student",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  /** Student ID, Teacher ID, or Admin ID */
  identifier: string;
  role: UserRole;
  department?: string | null;
  avatar?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export type CreateUserBodyRole =
  (typeof CreateUserBodyRole)[keyof typeof CreateUserBodyRole];

export const CreateUserBodyRole = {
  admin: "admin",
  teacher: "teacher",
  student: "student",
} as const;

export interface CreateUserBody {
  name: string;
  email: string;
  password: string;
  role: CreateUserBodyRole;
  department?: string;
}

export interface UpdateUserBody {
  name?: string;
  email?: string;
  department?: string;
}

export interface Class {
  id: number;
  name: string;
  code: string;
  teacherId: number;
  teacherName: string;
  studentCount: number;
  department?: string;
  createdAt: string;
}

export interface CreateClassBody {
  name: string;
  code: string;
  department?: string;
}

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  open: "open",
  closed: "closed",
  draft: "draft",
} as const;

export interface Project {
  id: number;
  title: string;
  description: string;
  classId: number;
  className: string;
  teacherId: number;
  teacherName: string;
  deadline: string;
  status: ProjectStatus;
  maxScore: number;
  submissionCount: number;
  createdAt: string;
}

export type CreateProjectBodyStatus =
  (typeof CreateProjectBodyStatus)[keyof typeof CreateProjectBodyStatus];

export const CreateProjectBodyStatus = {
  open: "open",
  closed: "closed",
  draft: "draft",
} as const;

export interface CreateProjectBody {
  title: string;
  description: string;
  classId: number;
  deadline: string;
  status: CreateProjectBodyStatus;
  maxScore: number;
}

export type UpdateProjectBodyStatus =
  (typeof UpdateProjectBodyStatus)[keyof typeof UpdateProjectBodyStatus];

export const UpdateProjectBodyStatus = {
  open: "open",
  closed: "closed",
  draft: "draft",
} as const;

export interface UpdateProjectBody {
  title?: string;
  description?: string;
  deadline?: string;
  status?: UpdateProjectBodyStatus;
  maxScore?: number;
}

export type SubmissionStatus =
  (typeof SubmissionStatus)[keyof typeof SubmissionStatus];

export const SubmissionStatus = {
  pending: "pending",
  reviewed: "reviewed",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface Submission {
  id: number;
  projectId: number;
  projectTitle: string;
  studentId: number;
  studentName: string;
  content: string;
  fileUrl?: string | null;
  status: SubmissionStatus;
  score?: number | null;
  feedback?: string | null;
  submittedAt: string;
  reviewedAt?: string | null;
}

export interface CreateSubmissionBody {
  projectId: number;
  content: string;
  fileUrl?: string;
}

export type UpdateSubmissionBodyStatus =
  (typeof UpdateSubmissionBodyStatus)[keyof typeof UpdateSubmissionBodyStatus];

export const UpdateSubmissionBodyStatus = {
  pending: "pending",
  reviewed: "reviewed",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface UpdateSubmissionBody {
  status?: UpdateSubmissionBodyStatus;
  score?: number;
  feedback?: string;
}

export interface DashboardStats {
  totalStudents: number;
  totalTeachers: number;
  totalProjects: number;
  totalSubmissions: number;
  pendingReviews: number;
  approvedSubmissions: number;
  activeClasses: number;
  submissionRate: number;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  submission: "submission",
  review: "review",
  project_created: "project_created",
  user_joined: "user_joined",
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  message: string;
  actorName: string;
  timestamp: string;
}

export interface TrendPoint {
  week: string;
  submissions: number;
  reviews: number;
}

export type ListUsersParams = {
  role?: ListUsersRole;
  search?: string;
};

export type ListUsersRole = (typeof ListUsersRole)[keyof typeof ListUsersRole];

export const ListUsersRole = {
  admin: "admin",
  teacher: "teacher",
  student: "student",
} as const;

export type ListProjectsParams = {
  classId?: number;
  status?: ListProjectsStatus;
};

export type ListProjectsStatus =
  (typeof ListProjectsStatus)[keyof typeof ListProjectsStatus];

export const ListProjectsStatus = {
  open: "open",
  closed: "closed",
  draft: "draft",
} as const;

export type ListSubmissionsParams = {
  projectId?: number;
  studentId?: number;
  status?: ListSubmissionsStatus;
};

export type ListSubmissionsStatus =
  (typeof ListSubmissionsStatus)[keyof typeof ListSubmissionsStatus];

export const ListSubmissionsStatus = {
  pending: "pending",
  reviewed: "reviewed",
  approved: "approved",
  rejected: "rejected",
} as const;