# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ProjTrack - Project Submission Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /classes:
    get:
      operationId: listClasses
      tags: [classes]
      summary: List classes
      responses:
        "200":
          description: List of classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Class"
    post:
      operationId: createClass
      tags: [classes]
      summary: Create class (teacher/admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClassBody"
      responses:
        "201":
          description: Class created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
  /classes/{classId}:
    get:
      operationId: getClass
      tags: [classes]
      summary: Get class by ID
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects
      parameters:
        - name: classId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [open, closed, draft]
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project (teacher/admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by ID
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    put:
      operationId: updateProject
      tags: [projects]
      summary: Update project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Project updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    LoginBody:
      type: object
      properties:
        identifier:
          type: string
          description: Email or ID
        password:
          type: string
        role:
          type: string
          enum: [admin, teacher, student]
      required:
        - identifier
        - password
        - role
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        identifier:
          type: string
          description: Student ID, Teacher ID, or Admin ID
        role:
          type: string
          enum: [admin, teacher, student]
        department:
          type: string
          nullable: true
        avatar:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - identifier
        - role
        - createdAt
...
    Class:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        teacherId:
          type: integer
        teacherName:
          type: string
        studentCount:
          type: integer
        department:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - code
        - teacherId
        - teacherName
        - studentCount
        - createdAt
    CreateClassBody:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        department:
          type: string
      required:
        - name
        - code
    Project:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        classId:
          type: integer
        className:
          type: string
        teacherId:
          type: integer
        teacherName:
          type: string
        deadline:
          type: string
          format: date-time
        status:
          type: string
          enum: [open, closed, draft]
        maxScore:
          type: integer
        submissionCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - classId
        - className
        - teacherId
        - teacherName
        - deadline
        - status
        - maxScore
        - submissionCount
        - createdAt
    CreateProjectBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        classId:
          type: integer
        deadline:
          type: string
          format: date-time
        status:
          type: string
          enum: [open, closed, draft]
        maxScore:
          type: integer
      required:
        - title
        - description
        - classId
        - deadline
        - status
        - maxScore
    UpdateProjectBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        deadline:
          type: string
          format: date-time
        status:
          type: string
          enum: [open, closed, draft]
        maxScore:
          type: integer
...